third_party/blink/web_tests/platform/generic/fast/forms/select/select-set-length-with-mutation-expected.txt

Tests setting the .length of an HTMLSelectElement with mutation listeners registered.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS sel.length is 200
PASS sel.length is 100
PASS sel.length is 180
PASS successfullyParsed is true

TEST COMPLETE
